+++ scripts/provision/rhel/custom/basic/A_base.sh
@@ -13,4 +13,8 @@ if [[ $(grep -E '^[A-Za-z ]*7\.' /etc/redhat-release) ]]; then
   systemctl disable firewalld
 fi
 
+if [[ -d /vagrant/shared/misc/tools/ && -d /etc/profile.d/ ]]; then
+  echo 'export PATH=${PATH}:/vagrant/shared/misc/tools/' > /etc/profile.d/vagrant_basic.sh
+fi
+
 echo "Finishing ${0}.."