+++ scripts/bootstrap/debian/2_packages.sh
@@ -4,7 +4,7 @@ echo '##############################################'
 echo "Starting ${0}.."
 set -x
 
-apt-get update
+apt-get -q update
 DEBIAN_FRONTEND=noninteractive apt-get -yV upgrade
 DEBIAN_FRONTEND=noninteractive apt-get -yV install apt-transport-https linux-headers-$(uname -r) bind9-host
 
@@ -13,7 +13,7 @@ cat <<EOF > /etc/init/refresh-apt.conf
 description "update package index"
 start on networking
 task
-exec /usr/bin/apt-get update
+exec /usr/bin/apt-get -q update
 EOF
 
 echo "Finishing ${0}.."