+++ scripts/provision/debian/custom/rex_install_git/1_base.sh
@@ -6,7 +6,7 @@ set -x
 
 echo "deb https://rex.linux-files.org/debian/ $(lsb_release -cs) rex" > /etc/apt/sources.list.d/rex.list
 wget -O - https://rex.linux-files.org/DPKG-GPG-KEY-REXIFY-REPO | apt-key add -
-apt-get update
+apt-get -q update
 
 # apt-cache depends rex | grep Depends | sed "s/.*ends:\ //" | tr '\n' ' '