+++ CHANGELOG.rst
@@ -3,6 +3,7 @@
 * update on demand only
 * change the location of some salt files, this will break backwards-compat
 * rename share/ to shared/
+* add /vagrant/shared/modules_ext as additional puppet mod dir
 
 v1.2.0 (2015-04-30)
 -------------------