+++ CHANGELOG.rst
@@ -6,6 +6,7 @@
 * rename share/ to shared/
 * add /vagrant/shared/modules_ext as additional puppet mod dir
 * optionally install hiera-eyaml
+* optionally add local tools to PATH
 
 v1.2.0 (2015-04-30)
 -------------------