+++ CHANGELOG.rst
@@ -8,6 +8,7 @@
 * optionally install hiera-eyaml
 * optionally add local tools to PATH
 * optionally specify puppet packages to install (e.g. to pin a version)
+* Order scripts by numbers instead of characters (#2)
 
 v1.2.0 (2015-04-30)
 -------------------